Splitted NavigationBindingsUtilities.
